third_party/WebKit/LayoutTests/touchadjustment/disabled-mobileviewport.html

<!DOCTYPE html>
<html>
<head>
    <title>Touch Adjustment : Disabled in the presence of a mobile-optimized viewport - bug 417534</title>
    <meta name="viewport" content="width=device-width, initial-scale=1">
    <script src="../resources/js-test.js"></script>
    <script src="resources/touchadjustment.js"></script>
    <style>
        #sandbox {
            position: absolute;
            left: 0px;
            top: 100px;
        }
        a {
            background-color: rgb(255,255,255);
        }
        a:active {
            background-color: rgb(0,0,255);
        }
    </style>
</head>

<body>

<div id="sandbox">
<p><a href="" id="link1">I</a> propose to consider <a href="" id="link2">the question</a>, "Can machines think?"<br>This should begin with definitions of the meaning of the terms "machine" and <a href="" id="link3">"think."</a></p>
</div>

<p id='description'></p>
<div id='console'></div>

<script>
    var element;
    var adjustedNode;
    // Set up shortcut access to elements
    var e = {};
    ['sandbox', 'link1', 'link2', 'link3'].forEach(function(a) {
        e[a] = document.getElementById(a);
    });
    
    var curElement;
    
    function isLinkActive()
    {
        // These need to match the background-color used above, after round-tripping.
        var defaultColor = "rgb(255, 255, 255)";
        var activeColor = "rgb(0, 0, 255)";
    
        color = window.getComputedStyle(curElement).backgroundColor;
        if (color == activeColor)
            return true;
        if (color != defaultColor)
            testFailed('Got unexpected backgroundColor: ' + color);
        return false;
    }
    
    function testTapDown(touchpoint, element, expectLinkActivated)
    {
        curElement = element;
        if (isLinkActive()) testFailed('Link unexpectedly active on entry');
        eventSender.gestureShowPress(touchpoint.x, touchpoint.y, touchpoint.width, touchpoint.height);
        if (expectLinkActivated) shouldBeTrue("isLinkActive()");
        else shouldBeFalse("isLinkActive()");
        eventSender.gestureTapCancel(touchpoint.x, touchpoint.y);
        if (isLinkActive()) testFailed('Link unexpectedly active on exit');
    }

    function testDirectTouch(element)
    {
        // Touch directly in the center of the element.
        var touchpoint = offsetTouchPoint(findAbsoluteBounds(element), 'center', 0, 20, 30);
        var willTapActivateLink = true;
        if (document.elementFromPoint(touchpoint.x, touchpoint.y) != element)
            testFailed('Direct touch ended up on some other element');
        testTapDown(touchpoint, element, willTapActivateLink);
    }

    function isDescendantOf(parent, child)
    {
        var n = child;
        while(n) {
            if(n==parent)
                return true;
            n = n.parentNode;
        }
        return false;
    }
    
    function testIndirectTouch(element)
    {
        // Touch just right of the element. With touch adjustment enabled, the
        // link element should not be activated.
        var touchpoint = offsetTouchPoint(findAbsoluteBounds(element), 'right', 10, 30, 20);
        var willTapActivateLink = false;
        if (isDescendantOf(element, document.elementFromPoint(touchpoint.x, touchpoint.y)))
            testFailed('Indirect touch ended up still on top of the element');
        testTapDown(touchpoint, element, willTapActivateLink);
    }

    function testDirectTouches()
    {
        debug('Testing direct hits.');
        testDirectTouch(e.link1);
        testDirectTouch(e.link2);
        testDirectTouch(e.link3);
    }

    function testIndirectTouches()
    {
        debug('Testing indirect hits.');
        testIndirectTouch(e.link1);
        testIndirectTouch(e.link2);
        testIndirectTouch(e.link3);
    }

    function runTests()
    {
        if (!window.eventSender) {
            debug('This test requires DumpRenderTree');
            return;
        }
        if (!window.eventSender.gestureShowPress) {
            debug('GestureShowPress not supported by this platform');
            return;
        }
        if (!window.internals) {
            debug('This test requires viewport support');
            return;
        }

        description('Tests that touch adjustment is disabled with a mobile-optimized viewport.');
        window.internals.settings.setViewportEnabled(true);
        testDirectTouches();
        testIndirectTouches();
        e.sandbox.style.display = 'none';
    }
    runTests();
</script>
</body>
</html>
