third_party/WebKit/LayoutTests/touchadjustment/disabled-mobileviewport-expected.txt

Tests that touch adjustment is disabled with a mobile-optimized viewport.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing direct hits.
PASS isLinkActive() is true
PASS isLinkActive() is true
PASS isLinkActive() is true
Testing indirect hits.
PASS isLinkActive() is false
PASS isLinkActive() is false
PASS isLinkActive() is false
PASS successfullyParsed is true

TEST COMPLETE
